+++ src/main/java/org/openhab/binding/openocean/internal/eep/Base/_VLDMessage.java
@@ -35,7 +35,7 @@ public abstract class _VLDMessage extends EEP {
     }
 
     @Override
-    protected boolean validateData(int[] bytes) {
+    protected boolean validateData(byte[] bytes) {
         return true;
     }
