+++ src/main/java/org/openhab/binding/openocean/internal/messages/ESP3PacketFactory.java
@@ -18,34 +18,34 @@ import org.openhab.binding.openocean.internal.messages.ESP3Packet.ESPPacketType;
  */
 public class ESP3PacketFactory {
 
-    public final static ESP3Packet CO_RD_VERSION = new CCMessage(1, 0, new int[] { 3 });
-    public final static ESP3Packet CO_RD_IDBASE = new CCMessage(1, 0, new int[] { 8 });
-    public final static ESP3Packet CO_RD_REPEATER = new CCMessage(1, 0, new int[] { 10 });
+    public final static ESP3Packet CO_RD_VERSION = new CCMessage(1, 0, new byte[] { 3 });
+    public final static ESP3Packet CO_RD_IDBASE = new CCMessage(1, 0, new byte[] { 8 });
+    public final static ESP3Packet CO_RD_REPEATER = new CCMessage(1, 0, new byte[] { 10 });
 
-    public static ESP3Packet CO_WR_IDBASE(int[] newId) {
-        return new CCMessage(5, 0, new int[] { 7, newId[0], newId[1], newId[2], newId[3] });
+    public static ESP3Packet CO_WR_IDBASE(byte[] newId) {
+        return new CCMessage(5, 0, new byte[] { 7, newId[0], newId[1], newId[2], newId[3] });
     }
 
     public static ESP3Packet CO_WR_REPEATER(StringType level) {
         switch (level.toString()) {
             case OpenOceanBindingConstants.REPEATERMODE_OFF:
-                return new CCMessage(3, 0, new int[] { 9, 0, 0 });
+                return new CCMessage(3, 0, new byte[] { 9, 0, 0 });
             case OpenOceanBindingConstants.REPEATERMODE_LEVEL_1:
-                return new CCMessage(3, 0, new int[] { 9, 1, 1 });
+                return new CCMessage(3, 0, new byte[] { 9, 1, 1 });
             default:
-                return new CCMessage(3, 0, new int[] { 9, 1, 2 });
+                return new CCMessage(3, 0, new byte[] { 9, 1, 2 });
         }
     }
 
     public static ESP3Packet CO_WR_SUBTEL(boolean enable) {
         if (enable) {
-            return new CCMessage(2, 0, new int[] { 17, 1 });
+            return new CCMessage(2, 0, new byte[] { 17, 1 });
         } else {
-            return new CCMessage(2, 0, new int[] { 17, 0 });
+            return new CCMessage(2, 0, new byte[] { 17, 0 });
         }
     }
 
-    public static ESP3Packet BuildPacket(int dataLength, int optionalDataLength, int packetType, int[] payload) {
+    public static ESP3Packet BuildPacket(int dataLength, int optionalDataLength, byte packetType, byte[] payload) {
         ESPPacketType type = ESPPacketType.getPacketType(packetType);
 
         switch (type) {
